chore: *looong* overdue code refactor.
