--- js/global.d.ts
@@ -1,5 +1,0 @@
-declare let process: {
-  env: {
-    BUGSNAG_KEY: string;
-  };
-};
