+++ js/modulepane/ModuleScoreBar.tsx
