+++ js/modulepane/ModuleTreeMap.tsx
@@ -1,7 +1,7 @@
 import { HierarchyRectangularNode, stratify, treemap } from 'd3-hierarchy';
 import React, { useEffect, useState } from 'react';
-import { BundlePhobiaData } from '../fetch_types.js';
 import $ from '../util/dom.js';
+import { BundlePhobiaData } from '../util/fetch_types.js';
 import human from '../util/human.js';
 
 export function ModuleTreeMap({
