+++ js/util/bugsnag.ts
@@ -2,6 +2,12 @@ import Bugsnag from '@bugsnag/js';
 import { version as appVersion } from '../../package.json';
 import HttpError from './HttpError.js';
 
+declare let process: {
+  env: {
+    BUGSNAG_KEY: string;
+  };
+};
+
 const bugsnag = Bugsnag.default.start({
   appVersion,
   apiKey: process.env.BUGSNAG_KEY,
